step3 : 피드백 반영(dashboard 객체 제거)
